test: move some tests from supi to pnpm
